Try to clean up gracefully if our app is forced to exit
Unfortunately, this does not work when the system is shutdown or rebooted. We may need to setup scripts to run as the run level changes to send signals to our scripts, but that's a bit involved for now. Instead, we just need to be sure to quit each app, or mainly pwm.py since we need it to reset to make sure the light and thruster are off.