Make each process use its own database
This is in the hopes that this will prevent corruption of the db as we say when all 5 processes were writing to a single db.

This also normalizes all sample rates to 1 second.

This fixes an issue in the pwm module where on/off values were not being logged during reset