First pass at handling ascent/descent axes
For now we only raise/lower the vertical thruster values by the most we can increase/decrease any of those values to reach the motor max/min. A later revision of this will start to flatten the curve between the thrusters, making the shape work towards a flat line, but this should work well for starters. I’m not even sure if the flat line approach just described will feel natural as it finds a compromise between the ascent/descent vector and whatever command is current in use. It may be better to lose some power in ascent/descent in order to preserve the feeling of pitch and roll.